--- bin/__init__.py
@@ -1,25 +1,0 @@
-#-*- coding: utf-8 -*-
-'''
-    Torrenter v2 plugin for XBMC/Kodi
-    Copyright (C) 2015 srg70, RussakHH, DiMartino
-'''
-
-from functions import *
-import xbmc, xbmcaddon
-import sys
-import os
-
-__settings__ = xbmcaddon.Addon(id='script.module.libtorrent')
-__version__ = __settings__.getAddonInfo('version')
-__plugin__ = __settings__.getAddonInfo('name') + " v." + __version__
-
-lm=LibraryManager(dest_path, platform)
-if not lm.check_exist():
-    ok=lm.download()
-    xbmc.sleep(2000)
-
-
-if __settings__.getSetting('plugin_name')!=__plugin__:
-    __settings__.setSetting('plugin_name', __plugin__)
-    lm.update()
-
